feat: created burger tree for refactor screen features
