+++ src/components/pages/main-page/header/burger/Burger.scss
@@ -1,5 +1,13 @@
+@import '../../../../../style/variables.scss';
+
 .Burger {
-  justify-content: flex-start;
+  position: absolute;
+  top: 0;
+  left: 0;
+
+  transition: all 0.5s ease;
+
+  display: none;
 
   .Burger-item {
     position: relative;
@@ -8,8 +16,6 @@
     margin-left: 20px;
     margin-top: 10px;
     cursor: pointer;
-
-    animation: burger-item-in 1s linear forwards;
 
     &::before {
       content: '';
@@ -51,6 +57,11 @@
         background-color: rgb(208, 204, 210);
       }
     }
+  }
+
+  @media (max-width: $mediaTablets) {
+    animation: burger-item-in 1s linear forwards;
+    display: block;
   }
 }
