+++ src/components/pages/main-page/header/nav-llist-item/NavListItem.scss
@@ -1,3 +1,5 @@
+@import '../../../../../style/variables.scss';
+
 .NavList__item {
   height: 48px;
   cursor: pointer;
@@ -28,6 +30,20 @@
     .NavList-link {
       text-shadow: none;
     }
+  }
+}
+
+.NavList-items--burger-will-hidden {
+  @media (max-width: $mediaTablets) {
+    animation: nav-item-hide 1s linear forwards;
+  }
+}
+
+@keyframes nav-item-hide {
+  100% {
+    transform: translateY(-50px);
+    opacity: 0;
+    display: none;
   }
 }
