+++ src/components/pages/main-page/projects/project-card/ProjectCard.scss
@@ -38,6 +38,7 @@
   padding: 5px;
   color: #fff;
   transition: 0.5s;
+  height: 400px;
 
   opacity: 0;
 
@@ -86,12 +87,10 @@
   grid-template-columns: repeat(2, 1fr);
   grid-gap: 20px;
 
-  @media (max-width: $mediaSmallDesktops) {
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    right: 0;
-  }
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  right: 0;
 }
 .Project-btn {
   display: flex;
