+++ src/store/nav/navSlice.js
@@ -3,17 +3,9 @@ import { createSlice } from '@reduxjs/toolkit';
 const navSlice = createSlice({
   name: 'nav',
   initialState: {
-    screenWidth: 768,
-    activeScreenBurgerMenu: false,
     activeBurgerMenu: false,
   },
   reducers: {
-    changeScreenWidth: (state, action) => {
-      state.screenWidth = action.payload;
-    },
-    checkActiveMenu: (state, action) => {
-      state.activeScreenBurgerMenu = action.payload;
-    },
     onChangeBurgerMenu: (state, action) => {
       state.activeBurgerMenu = action.payload;
     },
@@ -23,5 +15,4 @@ const navSlice = createSlice({
 const { actions, reducer } = navSlice;
 export default reducer;
 
-export const { changeScreenWidth, checkActiveMenu, onChangeBurgerMenu } =
-  actions;
+export const { onChangeBurgerMenu } = actions;
